renames api folder to services
